An error occurs when the token not contains the token_principal_attribute
This can happen if I need to support both access tokens received with the standard Authorization Code Flow and access tokens received with the Client Credentials Grant. In the former case, I set the "email" field in the user_provider_credential and token_principal_attribute settings, but in the latter case, there is no email address in the token, but I handle this with the cusom retrieve method.